requirements.txt
mysql-connector-python==9.1.0
tkcalendar==1.6.1